ACME: forbid specifying files in "acme_certificate key=...".
The ACME protocol assumes short-term certificates to ensure a regular
revalidation of ownership and to limit consequences of key compromise.
Allowing to request certificate renewals with the same key goes agains
the second goal and should not be supported without a good reason.

This was an oversight inherited from the proof-of-concept code and was
not supposed to be committed here.