ACME: assume "mailto:" scheme by default for contact URLs.
The only contact URL scheme required in RFC8555 is mailto:. While other
schemes are allowed and can be accepted by an ACME server, most of the
clients ignore that.

The "contact" directive parser is updated to prepend mailto: to any URLs
without a scheme. This can be resolved unambiguously, as ':' is not
allowed in the email address unquoted, and quotes are forbidden in the
URL scheme.