Net: add a simple HTTP client implementation.
The client is using NGINX connection infrastructure with async wrappers,
but the message parser is currently provided by "hyper".

There are several shortcomings in this approach, most importantly lack
of support for pre-resolved upstreams, keepalive or fine tuning of
connection buffering and timeouts.  There is a plan to provide a better
HTTP client implementation in one of the future releases of ngx-rust and
use it as a replacement.