Net: fix crash on HTTP connection to ACME server.
Plain HTTP connections are prohibited by the ACME specification, so we
did not have this scenario in our test automation and overlooked the
regression during the pre-release code cleanup. Nonetheless, some server
implementations allow such configuration and more importantly the HTTP
client code should be useful as an example for other modules.