ACME: preferred certificate chain support.
The new directive, "preferred_chain", allows selecting one of the
alternative certificate chains (RFC8555 § 7.4.2) by a Subject Common
Name that issued the topmost certificate in the chain.  If no suitable
chains are found, the default one will be used.

To fetch the alternative chains, we need a rather complicated parser for
RFC8288 Link header.  After considering the correctness, maintenance
status and dependency weight of the available implementations, I decided
to add our own instead, simplified as much as possible.