Reimplement DNS identifier validation.
The signature of ngx_http_validate_host has changed in NGINX 1.29.4,
breaking the build of the module.  Instead of continuing to abuse an
internal API and wrapping it with version check, this commit
reimplements the validation in a more suitable for the module way.

Notable differences from the previous behavior:

 - Minimal supported NGINX version is now aligned to the ngx crate: 1.22

 - The new logic mostly follows ngx_http_validate_host behavior as of
   NGINX 1.29.4, and may reject some previously accepted values.

 - Wildcard DNS identifiers are now permitted.  Some ACME servers allow
   validating wildcard identifiers with challenges other than dns-01.